requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
matplotlib==3.9.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.4
pytest==8.2.2
requests==2.32.3
scikit-learn==1.5.0
tensorflow==2.16.1
